+++ trys/app.py
@@ -1,6 +1,7 @@
 from flask import Flask, render_template, request, redirect, url_for, session
 import sqlite3
 import os
+import re
 print("👀 Current working dir:", os.getcwd())
 
 
@@ -11,6 +12,7 @@ app = Flask(__name__)
 app.secret_key = b'_5#y2L"F4Q8z\n\xec]/'
 
 
+# Connect to the SQLite database
 def get_db_connection():
     base_dir = os.path.dirname(os.path.abspath(__file__))
     db_path = os.path.join(base_dir, 'database.db')
@@ -20,10 +22,26 @@ def get_db_connection():
     return conn
 
 
+# If all are true, return true, else return false (readablitly improves)
+def is_strong_password(password):
+
+    if (
+        len(password) >= 8 and
+        re.search(r'[A-Z]', password) and
+        re.search(r'[a-z]', password) and
+        re.search(r'[0-9]', password) and
+        re.search(r'[!@#$%^&*(),.?":{}|<>]', password)
+    ):
+        return True
+    else:
+        return False
+
+
 @app.route('/')
 def index():
     if 'username' in session:   # check inside session dictionary, if there's 'username' data store
-        logged_msg = 'You are logged in...'
+        logged_msg = f'You are logged in... {session["username"]}.'
+
         return render_template('index.html', logged_msg=logged_msg)
     else:
         # redirect to login page if not logged in
@@ -61,6 +79,17 @@ def register():
     if request.method == 'POST':
         username = request.form['username']
         password = request.form['password']
+        confirm_password = request.form['confirm_password']
+
+        # Check if the passwords match
+        if password != confirm_password:
+            error = 'Passwords do not match.'
+            return render_template('register.html', error=error)
+
+        # Check if the password is strong
+        if not is_strong_password(password):
+            error = 'Password must be at least 8 characters long and contain uppercase, lowercase, digits, and special characters.'
+            return render_template('register.html', error=error)
 
         try:
             conn = get_db_connection()
@@ -71,7 +100,8 @@ def register():
             conn.commit()
             conn.close()
             return redirect(url_for('login3'))
-
+        # Handle the case where the username already exists
+        # username TEXT UNIQUE.. see this code in schema.sql, That UNIQUE makes sure no two rows can have the same username.
         except sqlite3.IntegrityError:
             error = 'Username already exists.'